refactor: move log into logging
